Refactor: 지원하기 페이지 변경된 디자인 반영 및 코드 리팩토링
Changes to be committed:
modified:   pages/apply/[platformName].tsx
modified:   src/components/apply/ApplyForm/ApplyForm.component.tsx
modified:   src/components/apply/ApplyForm/ApplyForm.styled.ts
modified:   src/components/apply/ApplyLayout/ApplyLayout.component.tsx
modified:   src/components/apply/ApplyLayout/ApplyLayout.styled.ts
